File re-organization and separating DB connection checks for simplicity
